split files for shadcn / mantine
